Make sure that ACL check are done for resource, not the ACL resource
Found a bug that uncovered a divergent behavior in NSS and ESS. Response header from NSS will serve Link rel="acl" for ACL resource to another ACL resource (that shouldn't exist, according to the spec), while ESS will serve Link rel"acl self" and point to the ACL resource itself.

This divergent behavior masked the bug when testing on NSS, which is why we haven't picked it up before.