Experiment with the latest auth
This is a preview of an upgrade to use the latest auth features, in
particular no longer relying on the ESS cookie to remain authenticated
after refresh.

This is mostly intended for additional testing to discover bugs and pain
points (of which I assume there will be ^^), and it can only be
considered for merging after thorough testing and release of a stable
version of the components.