Only fetching permissions from ACR + updating UI after changes in ACR
This introduces a weird UX-behavior, where the whole sidebar is updated, and the accordion for sharing is closed. This is quite annoying, but haven't found a workaround yet, and I wanted to share the other progress that I've made on this.

There is also a bug where the children container won't be updated with new policies added to the parent container - this is being looked into by the ESS team.