Re-enable legacy behavior
This adds a switch to configure the AcpAccessControlStrtegy in regular or legacy mode. In legacy mode, policies are assumed to be stored in external resources, which implies the policies content is savd before linking it to the ACR Access Control. In regular mode, policies are assumed to be inline of the ACR, which means they should be connected to the ACR's Access Control before their content is modified.

This also adds support for the legacy ACP API, based on Rules rather than Matchers.