Fixing tests
This (quite large) commit fixes the tests that have been broken by the introduction of the new behaviour. Namely, what it does is it duplicates a lot of the existing tests, and enforces that both the legacy and the latest behaviours are explicitly covered.

Enforce backward compatibility

This fixes issues preventing existing policies in 1.1 to be recognized.