Clean up useAllPermissions hook
The default states for consentPermisisons and acpPermissions were both
empty arrays. If they continued to be empty even after
getAllPermissionsforResource was called, the effect would not be called
because the parameters did not change.

I started by replacing the empty arrays in the useStates with null
instead of empty array, then realied that
normalizeConsentBasedPermissions can be called outside of the effect. So
instead, I removed most of the state, put that in a memo so that it
doesn't recalculate unless consentBasedPermissions changes, and set the
permissions based on the normalized permissions + consent permissions.