Remove the VC id from access request page
Podbrowser will be used as a consent management app by resource owners. This means that they will not be able to fetch the request VC to the issuer based on its IRI, because access control would enforce that only the requestor could access the VC. Instead, the VC is present encoded in the query strings of the request. This means that the VC IRI is useless to Podbrowser, and case in point it isn't used anywhere.

The current change makes it so that one can redirect a user to /privacy/consent/requests/ instead of /privacy/consent/requests/{encoded VC IRI}, which is simpler from a client app point of view.