refactor: Prevent rendering pages until session state is known and checked
This means that when components further down the tree render, the session is in a "known" state: either logged in or logged out, as the default state is unknown.
This does require a later commit to fully work (the returnTo URL wasn't respected by the login form, which needed a wider refactor).

Some of the changes here may make their way back into the Solid React SDK in the future.