fix: present grant_types in client to enable refresh_tokens
At some point, ESS made a change that meant it didn't by default issue refresh_tokens to clients that don't ask for them, as such, we need to explicitly ask for them. This only fixes deployments of Pod-Browser, but does not change local development, which needs the published localhost document updated.